Refactor packages for discord related data types
