+++ common/src/main/java/ac/grim/grimac/command/commands/GrimTestWebhook.java
@@ -6,7 +6,7 @@ import ac.grim.grimac.platform.api.manager.cloud.CloudCommandAdapter;
 import ac.grim.grimac.platform.api.sender.Sender;
 import ac.grim.grimac.utils.anticheat.LogUtil;
 import ac.grim.grimac.utils.anticheat.MessageUtil;
-import ac.grim.grimac.utils.webhook.WebhookMessage;
+import ac.grim.grimac.utils.webhook.discord.WebhookMessage;
 import org.incendo.cloud.CommandManager;
 import org.incendo.cloud.context.CommandContext;
 import org.jetbrains.annotations.NotNull;
