+++ common/src/main/java/ac/grim/grimac/manager/DiscordManager.java
@@ -10,7 +10,10 @@ import ac.grim.grimac.utils.anticheat.LogUtil;
 import ac.grim.grimac.utils.anticheat.MessageUtil;
 import ac.grim.grimac.utils.data.Pair;
 import ac.grim.grimac.utils.discord.CompiledDiscordTemplate;
-import ac.grim.grimac.utils.webhook.*;
+import ac.grim.grimac.utils.webhook.discord.Embed;
+import ac.grim.grimac.utils.webhook.discord.EmbedField;
+import ac.grim.grimac.utils.webhook.discord.EmbedFooter;
+import ac.grim.grimac.utils.webhook.discord.WebhookMessage;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
